+++ src/frontend/web_application/src/scenes/Signin.tsx
@@ -1,0 +1,3 @@
+import { SigninForm } from '../modules/user';
+
+export default SigninForm;