+++ src/frontend/web_application/src/scenes/Signup/index.jsx
@@ -1,3 +1,0 @@
-import { SignupForm } from '../../modules/user';
-
-export default SignupForm;