+++ src/frontend/web_application/src/services/username-utils/username-availability.ts
@@ -1,0 +1,9 @@
+import axios from 'axios';
+
+export const checkAvailability = async (username: string) => {
+  const response = await axios.get('/api/v2/username/isAvailable', {
+    params: { username },
+    headers: { 'X-Requested-With': 'XMLHttpRequest' },
+  });
+  return response.data.available;
+};